% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_initial_graph2.R
\name{build_initial_graph2}
\alias{build_initial_graph2}
\title{Building the initial graph}
\usage{
build_initial_graph2(target, true_dag)
}
\arguments{
\item{target}{a vector of nodes for which we want to conduct local pc on}

\item{true_dag}{a matrix containing all the information about the true dag}
}
\description{
This helps us to build the subgraphs for the local pc algorithm
We need the true DAG to provide the first and second order neighbors
of the set of target nodes
}
